move useToggleChat hook
